# Salus/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "medical API",
        "description": "API for medical advices and finding doctors.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/ai/last-advice": {
            "get": {
                "tags": [
                    "AI"
                ],
                "summary": "retrieve The latest advice",
                "operationId": "b4d6ef26219bcf54be46e931b418d87b",
                "responses": {
                    "200": {
                        "description": "List of user symptoms"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Symptoms not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/appointments": {
            "get": {
                "tags": [
                    "appointments"
                ],
                "summary": "Get all appointments for the authenticated user",
                "description": "Display a listing of the resource.",
                "operationId": "c68dab577f29c04c4422e6cb86fa36ce",
                "responses": {
                    "200": {
                        "description": "List of user appointments"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "appointments not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "appointments"
                ],
                "summary": "Update appointment",
                "description": "Update the specified resource in storage.",
                "operationId": "12cd7d983251abe69909c968ff208984",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointment_date",
                                    "doctor_id"
                                ],
                                "properties": {
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-03-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ],
                                        "example": "pending or confirmed"
                                    },
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "appointment Updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Store a newly created resource in storage.",
                "operationId": "0cfd7a3482605b745b634d591966d8bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointment_date",
                                    "doctor_id"
                                ],
                                "properties": {
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-03-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ],
                                        "example": "pending or confirmed"
                                    },
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Symptom created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/appointments/{appointment}": {
            "get": {
                "tags": [
                    "appointments"
                ],
                "summary": "Get one appointments for the authenticated user",
                "description": "Display the specified resource.",
                "operationId": "9b789941f10bf8dd3dea74aa7d6bd17c",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user appointment"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "appointment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "appointments"
                ],
                "summary": "Delete appointments for the authenticated user",
                "description": "Remove the specified resource from storage.",
                "operationId": "d2e4aef8da50e5d82f441ed65ea4700d",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user appointment"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "appointment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}